+++ MirrasENGINE/src/Core/Types/Integer.h
@@ -7,7 +7,7 @@ using int32 = int32_t;
 using int16 = int16_t;
 using int8  = int8_t;
 
-using uint64 = uint64_t;
+using uint64 = unsigned long long;
 using uint32 = uint32_t;
 using uint16 = uint16_t;
 using uint8  = uint8_t;
